Refactor test suite into organized categories for better maintainability
Reorganized the growing test suite from a flat structure into three logical
categories to improve organization, performance, and scalability:

- tests/tools/ (57 tests): Core LoadSet implementation tests - fast unit/integration tests
- tests/mcps/ (80 tests): MCP server mock tests - tool functionality without LLM calls
- tests/agents/ (31 tests): Agent integration tests - expensive LLM-based workflows

Key changes:
- Moved all test files to appropriate category directories
- Updated import paths to use tools.* prefix throughout
- Created conftest.py files for each directory with proper path resolution
- Updated pyproject.toml testpaths configuration for proper test discovery
- Enhanced CLAUDE.md with new test structure documentation and commands
- Extracted visual chart test from tools to agents category
- Maintained all existing functionality and test markers (@pytest.mark.expensive, @pytest.mark.visuals)

Benefits:
- Clear separation of fast vs expensive tests for better CI efficiency
- Logical grouping makes tests easier to find and maintain
- Scalable structure for adding new tests in appropriate categories
- Performance optimization: 137 fast tests run by default, 30 expensive tests run explicitly

All 168 tests verified working with new structure.
