+++ tests/agents/test_anthropic_mcp_http_integration.py
