+++ tests/agents/test_script_agent_integration.py
