+++ tests/mcps/conftest.py
@@ -1,8 +1,8 @@
 """
-Pytest configuration file for test discovery and import resolution.
+Pytest configuration file for MCP tests.
 
 This file helps pytest and VS Code properly discover tests and resolve imports
-from the tools directory.
+from the tools directory for tests in tests/mcps/.
 """
 
 import sys
@@ -12,7 +12,7 @@ import logfire
 from dotenv import load_dotenv
 
 # Add project root directory to Python path for imports
-project_root = Path(__file__).parent.parent
+project_root = Path(__file__).parent.parent.parent
 if str(project_root) not in sys.path:
     sys.path.insert(0, str(project_root))
