+++ tests/mcps/test_mcp_server_comparison.py
@@ -9,7 +9,7 @@ import tempfile
 import os
 from pathlib import Path
 
-from mcps.loads_mcp_server import create_mcp_server, reset_global_state
+from tools.mcps.loads_mcp_server import create_mcp_server, reset_global_state
 
 
 class TestMCPServerComparison:
