+++ tests/mcps/test_script_exec_mcp_server.py
@@ -14,11 +14,11 @@ from pathlib import Path
 import sys
 import time
 
-# Add tools directory to path
-tools_dir = Path(__file__).parent.parent / "tools"
-sys.path.insert(0, str(tools_dir))
+# Add project root to path
+project_root = Path(__file__).parent.parent.parent
+sys.path.insert(0, str(project_root))
 
-from mcps.script_exec_mcp_server import (
+from tools.mcps.script_exec_mcp_server import (
     ScriptExecutorMCPProvider,
     ExecutionResult,
     FileInfo,
